Add back the canonical urls/views
