+++ python-django/settings.py
@@ -76,6 +76,7 @@ WSGI_APPLICATION = 'wsgi.application'
 
 # Database
 # https://docs.djangoproject.com/en/2.2/ref/settings/#databases
+# WARNING: sqlite does not support serverless, use a remote DB
 
 DATABASES = {
     'default': {
